refactor(#292): split e2e staff-role tasks into per-task files composed of Interactions
Address PR #297 follow-up review comment: tasks/staff-roles-screen.ts held all
seven tasks plus helpers in one file. Now aligned with the acceptance-api
staff-role tasks/ layout, and introduces Serenity Interactions
(https://serenity-js.org/handbook/design/screenplay-pattern/#interactions) as
reusable command objects composed by tasks via Task.where:

- abilities/staff-portal-page.ts: page-access helpers shared by interactions
  and questions
- interactions/: OpenStaffRolesList, OpenCreateStaffRoleForm,
  OpenEditStaffRoleForm, FillStaffRoleForm, GrantFormPermission,
  SubmitStaffRoleForm, RecordListedStaffRoleNames, RecordBaselineStaffRoleCount
- tasks/: one file per task (create, ensure-exists, rename, grant-permission,
  view-list, open-screen), duplicated open/fill/submit logic removed

No behavior change: task names, step definitions, and questions surface are
unchanged; e2e suite passes 24/24.
